Exceptions/Errors and descriptions
Starting implement methods in cExceptionsAndErrors class, correction descriptions in other classes in project.

+++ Matrix_calculator/ccalculatorhandling.h
@@ -40,22 +40,22 @@ public:
     double d1, d2, d3;
 
     /*
-     * i - zmienna calkowita uzywana do wczytania
+     * short i - zmienna calkowita uzywana do wczytania
      * wykladnikow w potegowaniach liczb rzeczywistych
      * lub macierzy.
      */
-    int i;
+    short i;
 
     /*
-     * M1, M2, M3 - obiekty klasy cMatrix uzywane do
-     * wczytywania macierzy.
+     * cMatrix M1, M2, M3 - obiekty klasy cMatrix uzywane
+     * do wczytywania macierzy.
      */
     cMatrix M1, M2, M3;
 
 
 
     /*
-     * mMainHandling() - glowna metoda obslugi programu
+     * void mMainHandling() - glowna metoda obslugi programu
      * stanowiaca "rozgaleznik" miedzy operacjami gdzie
      * pierwszym argumentem jest liczba albo macierz.
      * PRE:
@@ -68,7 +68,7 @@ public:
 
 
     /*
-     * mChooseNumber() - glowna metoda obslugi programu
+     * void mChooseNumber() - glowna metoda obslugi programu
      * dla operacji w których pierwszym argumentem jest
      * liczba rzeczywista
      * PRE:
@@ -79,7 +79,7 @@ public:
     void mChooseNumber();
 
     /*
-     * mChooseNumberSum() - metoda obslugujaca dodawanie
+     * void mChooseNumberSum() - metoda obslugujaca dodawanie
      * i odejmowanie liczb rzeczywistych.
      * PRE:
      * - brak;
@@ -89,7 +89,7 @@ public:
     void mChooseNumberSum();
 
     /*
-     * mChooseNumberMultiplication() - metoda obslugujaca
+     * void mChooseNumberMultiplication() - metoda obslugujaca
      * mnozenie liczby przez liczbe rzeczywista albo
      * przez macierz.
      * PRE:
@@ -100,7 +100,7 @@ public:
     void mChooseNumberMultiplication();
 
     /*
-     * mChooseNumberOthers() - metoda obslugujaca dzielenie
+     * void mChooseNumberOthers() - metoda obslugujaca dzielenie
      * liczb rzeczywistych i potegowanie liczby rzeczywistej
      * do potegi calkowitej.
      * PRE:
@@ -114,7 +114,7 @@ public:
 
 
     /*
-     * mChooseMatrix() - metoda obslugujaca operacje gdzie
+     * void mChooseMatrix() - metoda obslugujaca operacje gdzie
      * pierwszym (i czasem jedynym) argumentem jest macierz.
      * PRE:
      * - brak;
@@ -124,7 +124,7 @@ public:
     void mChooseMatrix();
 
     /*
-     * mChooseMatrixSum() - metoda obslugujaca dodawanie
+     * void mChooseMatrixSum() - metoda obslugujaca dodawanie
      * i odejmowanie macierzy.
      * PRE:
      * - brak;
@@ -134,7 +134,7 @@ public:
     void mChooseMatrixSum();
 
     /*
-     * mChooseMatrixMultiplication() - metoda obslugujaca
+     * void mChooseMatrixMultiplication() - metoda obslugujaca
      * mnozenie macierzy przez liczbe rzeczywista albo
      * przez macierz.
      * PRE:
@@ -145,7 +145,7 @@ public:
     void mChooseMatrixMultiplication();
 
     /*
-     * mChooseMatrixExponentiation() - metoda obslugujaca
+     * void mChooseMatrixExponentiation() - metoda obslugujaca
      * potegowanie macierzy - w tym do potegi (-1) czyli
      * odwracanie macierzy.
      * PRE:
@@ -156,7 +156,7 @@ public:
     void mChooseMatrixExponentiation();
 
     /*
-     * mChooseMatrixProducts() - metoda obslugujaca
+     * void mChooseMatrixProducts() - metoda obslugujaca
      * iloczyny: skalarny i wektorowy.
      * PRE:
      * - brak;
@@ -166,7 +166,7 @@ public:
     void mChooseMatrixProducts();
 
     /*
-     * mChooseMatrixOthers() - metoda obslugujaca pozostale
+     * void mChooseMatrixOthers() - metoda obslugujaca pozostale
      * operacje na macierzach: dlugosc wektora, wyznacznik
      * macierzy, transponowanie macierzy.
      * PRE:
